Extraction des fonctionnalités de la popup des favoris dans un fichier dédié
